refactor: 중복된 객체 삭제
